Move our test data to the www directory
Cordova fails to include files outside of www in actual apps built and
deployed to devices.
Accesses to the sample data worked in simulated browsing, but not in
actual testing on Android devices.

Move the test data into the www directory to fix this.

Additionally, fix some case-sensitivity issues in the naming which may
have been masked on Windows and hid this problem in previous testing.